Python
- fix routes
- added images in quiz
- add missions in python
