+++ Assets/Models/UI/Quiz/QuizImages/Python/Comments.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d09c46b808adc148a5030dca3b3f97d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
