+++ Assets/Models/UI/Quiz/QuizImages/Python/List.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 979c9ac198503bf4d9724b2c2df9467f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
