+++ Assets/Models/UI/Quiz/QuizImages/Python/Operators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aac0abeb01b8d5443bb301bb8b316fe4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
