+++ Assets/Models/UI/Quiz/QuizImages/Python/PrintingValues.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e13d3e725ddc4cd4fbdb78db4a3279b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
