+++ Assets/Models/UI/Quiz/QuizImages/Python/Variables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12c31da3896f33c4bab0a3224fde4ae3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
