+++ Assets/Scripts/LayoutRefresher.cs
@@ -1,0 +1,32 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+public class LayoutRefresher : Singleton<LayoutRefresher>
+{
+    public void RefreshContentFitter(RectTransform transform)
+    {
+        if (transform == null || !transform.gameObject.activeSelf)
+        {
+            return;
+        }
+
+        foreach (RectTransform child in transform)
+        {
+            RefreshContentFitter(child);
+        }
+
+        var layoutGroup = transform.GetComponent<LayoutGroup>();
+        var contentSizeFitter = transform.GetComponent<ContentSizeFitter>();
+        if (layoutGroup != null)
+        {
+            layoutGroup.SetLayoutHorizontal();
+            layoutGroup.SetLayoutVertical();
+        }
+
+        if (contentSizeFitter != null)
+        {
+            LayoutRebuilder.ForceRebuildLayoutImmediate(transform);
+        }
+    }
+}
