+++ Assets/Scripts/LayoutRefresher.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d005e57fac69144d89df9b8ac379da5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
