+++ Assets/Scripts/Lesson/Instructions.cs
@@ -9,30 +9,20 @@ public class Instructions : MonoBehaviour
     GameObject parentGameObject;
     [SerializeField] string instruction;
     string _playerName;
-    // [SerializeField] Sprite syntaxImage;
-    // public Image img;
     void Start()
     {
         _playerName = DataManager.GetPlayerName();
-        // ChangeImage();
+        LayoutRefresher.Instance.RefreshContentFitter((RectTransform) transform);
     }
     void OnEnable()
     {
         isActive = true;
         Instruction();
-        RefreshContentFitter((RectTransform) transform);
     }
-
     void OnDisable()
     {
         isActive = false;
     }
-    // public void ChangeImage ()
-    // {  
-    //     if (img != null) img = GetComponent<Image>();
-    //     img.sprite = syntaxImage;
-
-    // }
     public void Instruction()
     {
         if (isActive)
@@ -44,30 +34,5 @@ public class Instructions : MonoBehaviour
     {
         NPCDialogue.Instance.AddDialogue(message, _playerName);
         NPCDialogue.Instance.ShowDialogue();
-    }
-    public void RefreshContentFitter(RectTransform transform)
-    {
-        if (transform == null || !transform.gameObject.activeSelf)
-        {
-            return;
-        }
-
-        foreach (RectTransform child in transform)
-        {
-            RefreshContentFitter(child);
-        }
-
-        var layoutGroup = transform.GetComponent<LayoutGroup>();
-        var contentSizeFitter = transform.GetComponent<ContentSizeFitter>();
-        if (layoutGroup != null)
-        {
-            layoutGroup.SetLayoutHorizontal();
-            layoutGroup.SetLayoutVertical();
-        }
-
-        if (contentSizeFitter != null)
-        {
-            LayoutRebuilder.ForceRebuildLayoutImmediate(transform);
-        }
     }
 }
